Allows alternate image names for testing
Adjusts build.sh and ocm-container.sh to allow optional arguments for the container image name to use, to make testing easier.  This allows for development and testing of an ocm-container image without breaking an image you might be using in production (for example, if you're currently shift Primary...theoretically...).

Usage:

```
./build
./ocm-container

./build ocm-container:testing
./ocm-container ocm-container:testing
```

Signed-off-by: Christopher Collins <collins.christopher@gmail.com>